32bit->64bit 수정 및 세션 패킷 버퍼 동적 할당, 패킷 수신용 스레드 변수 UPROPERTY 설정

+++ Source/TestProject/Packet.cpp
@@ -65,7 +65,7 @@ char* Packet::GetBuffer() const
 	return buffer;
 }
 
-ULONG Packet::GetPacketSize() const
+size_t Packet::GetPacketSize() const
 {
 	return packetSize;
 }
